+++ src/Structures/Self.js
@@ -16,7 +16,7 @@ class Self {
 		* The main client instance.
 		* @type {Client}
 		*/
-		Object.defineProperty(this, 'client', { value: client});
+	 this.client = client;
 		
 		/**
 		* The raw object data.